+++ src/fourlang/longman_parser.py
@@ -5,36 +5,9 @@ import json
 import re
 import sys
 
+from xml_parser import XMLParser
+
 assert json  # silence pyflakes
-
-class XMLParser():
-
-    @staticmethod
-    def section_pattern(tag):
-        pattern_string = "<{0}>(.*?)</{0}>".format(tag)
-        return re.compile(pattern_string, re.S)
-
-    @staticmethod
-    def tag_pattern(tag):
-        pattern_string = "</?{0}>".format(tag)
-        return re.compile(pattern_string, re.S)
-
-    @staticmethod
-    def iter_sections(tag, text):
-        return XMLParser.section_pattern(tag).findall(text)
-
-    @staticmethod
-    def get_section(tag, text):
-        match_obj = XMLParser.section_pattern(tag).search(text)
-        return None if match_obj is None else match_obj.group(1)
-
-    @staticmethod
-    def remove_sections(tag, text):
-        return XMLParser.section_pattern(tag).sub("", text)
-
-    @staticmethod
-    def remove_tags(tag, text):
-        return XMLParser.tag_pattern(tag).sub("", text)
 
 class LongmanParser(XMLParser):
 
@@ -116,4 +89,3 @@ class LongmanParser(XMLParser):
 
 if __name__ == "__main__":
     LongmanParser.print_defs(LongmanParser.parse_file(sys.argv[1]))
-    #json.dump(LongmanParser.parse_file(sys.argv[1]), sys.stdout)
